+++ NEWS.md
@@ -1,5 +1,6 @@
 # surveydown (development version)
 
+- Require that the survey file be named `"survey.qmd"`.
 - Added `sd_include_folder()` function so users can add a folder to the shiny resource path.
 - Automatically include the `images`, `js`, `css`, and `www` folders as well as folders to quarto files to the shiny resource path when the package loads.
 - Heavily modified how `sd_next()` works to improve page navigation and ensure that each `sd_next()` button has a unique id based on the current page.